I wrote some configuration management code to allow everyone to maintain local variations in things like directory structure without needing to alter the project code for everyone. Basically, you modify your own local copy of ohsudream-config.R. My 'training_data_stats.R' script sources this file and even prints a little error message if you don't have it. The 'ohsudream-config-example.R' file is, however, part of the git repo and is what the script will print if you don't have your own local version.
The 'training_data_stats.R' script is just a renamed and modified version of my original test script. Essentially it collects some stats on the files. It looks at things like stdev, median, and mean both across cell lines and across genes. My thinking was that we might look for genes that showed high or low variation with regards to cell lines as well as cell lines that showed high or low variation vis. a vis. genes. I also hoped to look at variations between median and means as a measure of skewness though I'm sure there are more sophisticated things we could do. Hope this is clear.